% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaPlot.R
\name{AlphaPlot}
\alias{AlphaPlot}
\title{AlphaPlot (a cross plot of live-dead offsets in alpha diversity and evenness)}
\usage{
AlphaPlot(x, colpt = "gray", colci = "gray", bgpt = "white",
  cf.bars = TRUE, cex = 0.8, pchpt = 21, colmean = "black",
  col.gp = NULL, addlegend = TRUE)
}
\arguments{
\item{x}{An object (a list) returned by FidelityDiv function}

\item{colpt}{A character string or number defining color of symbols for individual sites
(default = 'gray')}

\item{colci}{A character string or number defining color of confidence bars for individual
sites (default = 'gray')}

\item{bgpt}{A character string or number defining color of symbols for individual
sites (default = 'white')}

\item{cf.bars}{Logical (default=TRUE) to indicate if confidence bars should be plotted}

\item{cex}{a numerical value (default=0.8) defining expansion factor for symbols}

\item{pchpt}{A number defining symbol type (default = 21) (filled symbols (21-25) should be used)}

\item{colmean}{A character string or number defining color of bars representing
mean estimates for all sites (default = 'white')}

\item{col.gp}{A character string or number defining colors for group (default = 1:#levels)
for symbols and bars for sites (applicable when 'gp' factor is provided). The number of colors
must match number of levels in a grouping factor}

\item{addlegend}{Logical (default=TRUE): add legend
Note: The legend is plotted only when 'gp' factor provided}
}
\value{
A single bivariate plot produce by plot function
}
\description{
AlphaPlot function generates a bivariate plot of an offset in alpha diversity (DELTA S)
on x axis versus an offset in evenness (DELTA PIE) for all live-dead pairwise comparisons.
}
\details{
AlphaPlot function uses the output of FidelityDiv function to produce
a cross plot visualizing differences in estimates of alpha diversty and evenness between
pairs of sympatric live and dead samples. In its default form, an approch proposed by
Olszewski and Kidwell (200X) is used. The offset in alpha diversity is measured as the difference
between natural logarithms of sample standardized richness log(R-dead) - log(R-live). The offset
in evenness is measured as the difference between Hurlbert's PIE (PIE-dead - PIE-live).
Confidence bars depitcting user-specified confidence intervals are also plotted.
If a grouping factor is provided, symbols and bars are color-coded by levels and
group means are plotted in addition to grand mean for all sites.
}
\examples{

my.fid1 <- FidelityDiv(FidData$live, FidData$dead, n.filters=30)
AlphaPlot(my.fid1)
my.fid2 <- FidelityDiv(FidData$live, FidData$dead, FidData$habitat, n.filters=50, iter=1000, CI=0.95)
AlphaPlot(my.fid2, col.gp=c('forestgreen', 'coral1'), bgpt='beige')

}
