% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelityEst.R
\name{FidelityEst}
\alias{FidelityEst}
\title{FidelityEst computes live-dead fidelity measures}
\usage{
FidelityEst(live, dead, gp = NULL, cor.measure = "spearman",
  sim.measure = "chao", report = FALSE)
}
\arguments{
\item{live}{A species/taxon abundance matrix with counts of live-collected specimens
(rows=samples, columns=species/taxa)}

\item{dead}{A matrix with counts of dead specimens (rows=samples, columns=species/taxa)
(dimensions of 'live' and 'dead' must match)}

\item{gp}{An optional factor, with two or more levels, defining sample groups
(the length of gp must equal number of rows in live and dead)}

\item{cor.measure}{defines correlation coefficient (stats function 'vignettecor') used to measure pairwise
live-dead correlations (default='spearman').}

\item{sim.measure}{A measure of similarity to measure live-dead agreement (default='Chao')}

\item{report}{Print compliance report from function FidelitySummary}
}
\value{
A list containing the following components:
  \item{x}{The values of the Spearman coefficients \emph{rho} for each live-dead comparison}
}
\description{
FidelityEst computes common live-dead fidelity measures by comparing two matching
matrices (live and dead) with species/taxon abundance data.
In the case of datastes representing more than one sample,
the function returns also means. If 'gp' factor is provided to define sample groups
means for groups are returned.
}
\examples{

data(FidData)
FidelityEst(live=FidData$live, dead=FidData$dead, gp=FidData$habitat)

}
