% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDPlot.R
\name{LDPlot}
\alias{LDPlot}
\title{A Comparative Live-Dead Barplot}
\usage{
LDPlot(live, dead, tax.names, toplimit = 10, barwidth = 150/toplimit,
  col1 = "black", col2 = "gray", arr.col = "red", arr.lty = 1,
  cex.axis = 0.7, tck = -0.02, cex.lab = 0.8, cex.names = 0.95,
  cex.label = 1)
}
\arguments{
\item{live}{A vector with integer counts of live specimens}

\item{dead}{A vector with integer counts of dead specimens}

\item{tax.names}{A vector with a list of names of taxa (or other units
used as variables)}

\item{toplimit}{A numerical value (default = 10) defining the number of top
species (or other variables) to be plotted}

\item{barwidth}{A numerical value (default = 250 / toplimit) defining bar width
(or more precisely the thicknes of graphics::lines used to represent bars)}

\item{col1}{A character string (default = 'black') defining the color of bars
for taxa (or other variables) shared by live and dead data}

\item{col2}{A character string (default = 'gray') defining the color of bars
for taxa (or other variables) unique to either live or dead data}

\item{arr.col}{A character string (default = 'red') defining the color of arrows}

\item{arr.lty}{An integer or character string specifying type of line used for
arrows (default = 1) (this value is passed on to the graphical parameter 'lty')}

\item{cex.axis}{A numerical value (default = 0.7).The magnification to be used for
x axis annotation relative to the current setting of cex.}

\item{tck}{A numerical value (default  = -0.02) defining length of tick marks
for x axis}

\item{cex.lab}{A numerical value (default = 0.8) defining font size for
x-axis label}

\item{cex.names}{A numerical value (default = 0.95) defining font size
for names of taxa/variables corresponding to individual bars on the barplot}

\item{cex.label}{A numerical value (defult = 1) defining font size for 'Live'
and 'Dead' titles placed above the barplots}
}
\value{
A single plot.
}
\description{
LDPlot function generates a comparative live-dead plot of the most frequent taxa
(or other units/variables describing samples) to highlight conguences and discordances
in relative abundance and rank order of most abundant taxa/variables.
}
\details{
LDPlot function is designed to produce a plot with two barplots comparing
relative abundances of the most common taxa in live and dead datasets. This plot
is suitable for single live-dead comparisons only (single samples,
sets of samples or datasets).

NOTE: This function utilizes graphics::plot function, including some of its common
graphic arguments. It allows for quick exploratory plots and should be readily
editable to derive more customized plots.
}
\examples{

temp.par <- par(mar=c(3,5,2,5))
LDPlot(live=colSums(FidData$live), dead=colSums(FidData$dead),
       tax.names=colnames(FidData$live), toplimit=15, col1 = 'green',
       col2 = 'red4', arr.col = 'black', arr.lty = 3)
par(temp.par)

}
