% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleoFidelity.R
\docType{package}
\name{PaleoFidelity-package}
\alias{PaleoFidelity-package}
\alias{PaleoFidelity}
\title{PaleoFidelity: Statistical assessment of paleontological fidelity (live-dead community agreement)}
\description{
PaleoFidelity package provides common measures of paleontological fidelity (live-dead community
agreement), including compositional agreement, alpha diversity, and beta diversity.
}
\details{
PaleoFidelity offers permutation tests for one-sample and multi-sample tests, and
bootstrapped confidence intervals for fidelity statistics.

Several plot functions are included to visualize outcomes of fidelity analyses.
}
\examples{
# Version and citation
packageVersion('PaleoFidelity')
citation('PaleoFidelity')

# Check if data are compliant and generate basic data summary
\dontrun{require(vegan)
data(dune)
FidelitySummary(as.matrix(dune), as.matrix(dune[sample(1:nrow(dune)),]), report=T)

# Compute measures of compositional fidelity
out1 <- FidelityEst(as.matrix(dune), as.matrix(dune[sample(1:nrow(dune)),]))
plot(out1)}
}
\author{
Michal Kowalewski \email{kowalewski@ufl.edu}
}
