% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleoFidelity.R
\docType{package}
\name{PaleoFidelity}
\alias{PaleoFidelity}
\title{PaleoFidelity: Measuring paleontological fidelity (live-dead community agreement)}
\description{
PaleoFidelity package provides common measures of paleontological fidelity
used to assess live-dead community agreement. These measures encompass
multiple aspects of fidelity, including faunal composition and alpha diversity.
}
\details{
Plot functions to visualize fidelity patterns are also provided.
}
\examples{
# Version and citation
packageVersion('PaleoFidelity')
citation('PaleoFidelity')

# Check if data are compliant and generate basic data summary
data(FidData)
FidelitySummary(live=FidData$live, dead=FidData$dead, gp=FidData$habitat, report=TRUE)

# Returns estimates of compositional fidelity
FidelityEst(live=FidData$live, dead=FidData$dead, gp=FidData$habitat, iter=49)


}
\author{
Michal Kowalewski \email{kowalewski@ufl.edu}
}
