% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaPlot.R
\name{AlphaPlot}
\alias{AlphaPlot}
\title{A bivariate plot of live-dead offsets in alpha diversity and evenness}
\usage{
AlphaPlot(x, col = "gray", col.bar = "gray", bgpt = "white",
  cf.bars = TRUE, cex = 0.8, legend.cex = 1, pch = 21,
  colmean = "black", col.gp = NULL, addlegend = TRUE)
}
\arguments{
\item{x}{An object (a list) returned by FidelityDiv function}

\item{col}{A character string or number defining color of symbols for individual sites
(default = 'gray'). This argument applies only if a grouping factor 'gp' is not provided.}

\item{col.bar}{A character string or a number defining color of confidence bars for individual
sites (default = 'gray'). This argument applies only if a grouping factor 'gp' is not provided.}

\item{bgpt}{A character string or number defining background color of symbols for individual
sites (default = 'white').}

\item{cf.bars}{Logical (default=TRUE) to indicate if confidence bars should be plotted.}

\item{cex}{A numerical value (default=0.8) defining expansion factor for symbols.}

\item{legend.cex}{A numerical value (default = 1) defining text size for legend.}

\item{pch}{An integer defining symbol type (default=21). Filled symbols (21-25) should be used.}

\item{colmean}{A character string or number defining color of bars representing
mean estimates for all sites (default = 'white')}

\item{col.gp}{A character string or number defining colors for group (default = 1:#levels)
for symbols and bars for sites (applicable when 'gp' factor is provided). The number of colors
must match number of levels in a grouping factor}

\item{addlegend}{Logical (default=TRUE): adds legend to the plot, if 'gp' factor is provided.}
}
\value{
A single bivariate plot produce by plot function.
}
\description{
AlphaPlot function generates a bivariate plot of offsets in alpha diversity (DELTA S; x-axis)
and evenness (DELTA PIE; y-axis) for within-site pairwise comparisons of live and dead samples.
}
\details{
AlphaPlot function uses the output of FidelityDiv function to produce a crossplot
visualizing live-dead differences in alpha diversity and evenness. Confidence bars around
DELTA S and DELTA PIE estimates are also plotted. If a grouping factor is provided, symbols
and bars are color-coded by levels and group means are plotted in addition to grand mean
for all sites.
}
\examples{

my.fid1 <- FidelityDiv(FidData$live, FidData$dead, n.filters=30)
AlphaPlot(my.fid1)
my.fid2 <- FidelityDiv(FidData$live, FidData$dead, FidData$habitat,
                       n.filters=50, iter=1000, CI=0.95)
AlphaPlot(my.fid2, col.gp=c('forestgreen', 'coral1'), bgpt='beige')

}
