% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaPlot.R
\name{AlphaPlot}
\alias{AlphaPlot}
\title{AlphaPlot (outputs a cross plot of live-dead offsets in alpha diversity and evenness)}
\usage{
AlphaPlot(x, CI = 0.5, CImean = 0.99, colpt = "gray",
  colci = "gray", bgpt = "white", pchpt = 21, colmean = "black",
  colpt.gp = NULL, colci.gp = NULL, addlegend = TRUE)
}
\arguments{
\item{x}{An object (a list) returned by FidelityDiv function}

\item{CI}{A numerical value (default = 0.5) defining confidence bars for individual sites.
Note: 0.5 - plots bars represnting interuqartile range, 0.95 - plots 95% confidence bars, etc.
Confidence bars are estimated as percentiles of subsampled estimates of Delta S and Delta PIE.}

\item{CImean}{A numerical value (default = 0.99) defining confidence bars for means of all sites
or groups of sites (if 'gp' factor was provided). Note: 0.5 - plots bars represnting interuqartile
range, 0.95 - plots 95% confidence bars, etc. Confidence bars are estimated as percentiles of
subsampled estimates of Delta S and Delta PIE.}

\item{colpt}{A character string or number defining color of symbols for individual sites
(default = 'gray')}

\item{colci}{A character string or number defining color of confidence bars for individual
sites (default = 'gray')}

\item{bgpt}{A character string or number defining color of symbols for individual
sites (default = 'white')}

\item{pchpt}{A number defining symbol type (default = 21)
Note: filled symbol numbers (21-25) should be used}

\item{colmean}{A character string or number defining color of bars representing
mean estimates for all sites (default = 'white')}

\item{colpt.gp}{A character string or number defining colors by group (default = 1:#levels)
for symbols for sites (applicable when 'gp' factor is provided). Custom colors
can be provided, but must match number of levels in a grouping factor}

\item{colci.gp}{A character string or number defining colors by group (default = 1:#levels)
for confidence bars for sites (applicable when 'gp' factor is provided). Custom colors
can be provided, but must match number of levels in a grouping factor}

\item{addlegend}{Logical (default=TRUE): add legend (activated only when 'gp' factor provided)}
}
\value{
A single bivariate plot produce by plot function
}
\description{
AlphaPlot function generates a bivariate plot of an offset in alpha diversity (DELTA S)
on x axis versus an offset in evenness (DELTA PIE) for all live-dead pairwise comparisons.
}
\details{
AlphaPlot function uses the output of FidelityDiv function to produce
a cross plot visualizing differences in estimates of alpha diversty and evenness between
pairs of sympatric live and dead samples. In its default form, an approch proposed by
Olszewski and Kidwell (200X) is used. The offset in alpha diversity is measured as the difference
between natural logarithms of sample standardized richness log(R-dead) - log(R-live). The offset
in evenness is measured as the difference between Hurlbert's PIE (PIE-dead - PIE-live).
Confidence bars depitcting user-specified confidence intervals are also plotted.
If a grouping factor is provided, symbols and bars are color-coded by levels and
group means are plotted in addition to grand mean for all sites.
}
\examples{

AlphaPlot(FidelityDiv(FidData$live, FidData$dead, FidData$habitat, n.filters=30, t.filters=1))

}
