% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelityDiv.R
\name{FidelityDiv}
\alias{FidelityDiv}
\title{Live-Dead differences in alpha diversity and evenness}
\usage{
FidelityDiv(live, dead, gp = NULL, report = FALSE, n.filters = 0,
  t.filters = 1, iter = 100, CI = 0.5, CImean = 0.99,
  outdata = FALSE)
}
\arguments{
\item{live}{A matrix with counts of live-collected specimens (rows=sites, columns=taxa).
Dimensions of 'live' and 'dead' matrices must match exactely.}

\item{dead}{A matrix with counts of dead-collected specimens (rows=sites, columns=taxa).
Dimensions of 'live' and 'dead' matrices must match exactely.}

\item{gp}{An optional univariate factor defining groups of sites. The length of gp must
equal number of rows of 'live' and 'dead' matrices.}

\item{report}{Logical (default=FALSE) to print compliance report from function FidelitySummary}

\item{n.filters}{An integer used to filter out small samples (default n.filters=0, all samples kept)}

\item{t.filters}{An integer used to filter out rare taxa (default t.filters=1, taxa >= 1 occurrence kept)}

\item{iter}{An integer defining number of resampling iteration (default iter=100)}

\item{CI}{A numerical value (default = 0.5) defining confidence bars for individual sites.
Note: 0.5 - plots bars represnting inter-quartile ranges, 0.95 - plots 95% confidence bars, etc.
Confidence bars are estimated as percentiles of subsampled estimates of Delta S and Delta PIE.}

\item{CImean}{A numerical value (default = 0.99) defining confidence bars for means of all sites
or groups of sites (if 'gp' factor was provided). Note: 0.5 - plots bars represnting inter-quartile
range, 0.95 - plots 95% confidence bars, etc. Confidence bars are estimated as percentiles of
subsampled estimates of Delta S and Delta PIE.}

\item{outdata}{Logical (default = FALSE) to determine if data files should be included in the output}
}
\value{
A list containing the following components:
  \item{live}{The post-processed version of 'live' data matrix used in all analyses}
  \item{dead}{The post-processed version of 'dead' data matrix used in all analyses}
  \item{gp}{The post-processed version of 'gp' factor, when provided}
  \item{x}{DELTA S values for each live-dead comparisons (site-level differences in sample standardized
           species richness)}
  \item{y}{DELTA PIE values for each live-dead comparisons (site-level differences in evenness
           estimated as Hurlbert's PIE)}
  \item{xmean}{Grand mean of Delta S}
  \item{ymean}{Grand mean of Delta PIE}
  \item{xmean}{Grand mean of Delta S}
  \item{xgp}{Group means of Delta S (when 'gp' factor provided)}
  \item{ygp}{Group means of Delta PIE (when 'gp' factor provided)}
  \item{p.values}{p.values for Null H: Delta.S.p=0 and Delta.PIE.p=0}
  \item{p.gps}{per-group p.values for Null H: Delta.S.p=0 and Delta.PIE.p=0}
}
\description{
FidelityDiv provides estimates of differences in diversity between pairs of live and dead samples
using community abundance data. In the case of datastes representing more than one site, the
function returns also means of differences. If 'gp' factor is provided to aggregate sets of
sites/samples, means for groups are returned as well.
}
\details{
FidelityDiv assess live-dead offsets in evenness/diversity between pairs of sympatric
live and dead samples using a bivariate approch as described in  Olszewski and Kidwell (2007).
The estimates of (1) alpha diversity and (2) evenness are returned as 2 separate objects.

(1) x - Live-dead offsets in alpha diversity for individual sites. The difference is measured
as the difference between natural logarithms of sample-standardized species richness of
sympatric dead and live samples:

DELTA S = ln(S-dead) - log(S-live)
A negative value indicates that the sample standardized alpha diversity of a live sample
exceeds alpha diversity of the sympatric dead sample (and vice versa). Confidence intervals
and p.values for Null H: Delta S = 0 are also reported.

(2) y - Live-dead offsets in evenness for individual sites. The difference is measured as the
difference between sample-standardized estimates of Hurlbert's PIE for live and dead samples:

DELTA PIE = PIE(DEAD) - PIE(LIVE)

A negative value indicates that evenness of live samples exceeds evenness of the sympatric
dead sample (and vice versa). Confidence intervals and p.values for Null H: Delta PIE = 0
are also reported.
}
\examples{

FidelityDiv(FidData$live, FidData$dead, n.filters=30)
my.fid <- FidelityDiv(FidData$live, FidData$dead, FidData$habitat, n.filters=50, iter=1000, CI=0.95)
my.fid$x # site-level estimates of Delta S with 95\% CIs and p values
my.fid$p.gps # p values for means of groups

}
\references{
Olszewski, T.D., and Kidwell, S.M., 2007, The preservational fidelity of evenness
            in molluscan death assemblages. Paleobiology 33: 1:23.
}
