% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelityDiv.R
\name{FidelityDiv}
\alias{FidelityDiv}
\title{FidelityDiv estimates live-dead differences in alpha and beta diversity}
\usage{
FidelityDiv(live, dead, gp = NULL, report = FALSE, n.filters = 0,
  t.filters = 1, iter = 100)
}
\arguments{
\item{live}{A species/taxon abundance matrix with counts of live-collected specimens
(rows=samples, columns=species/taxa)}

\item{dead}{A matrix with counts of dead specimens (rows=samples, columns=species/taxa)
(dimensions of 'live' and 'dead' must match)}

\item{gp}{An optional factor, with two or more levels, defining sample groups
(the length of gp must equal number of rows in live and dead)}

\item{report}{Logical (default=FALSE) to print compliance report from function FidelitySummary}

\item{n.filters}{An integer used to filter out small samples (default n.filters=0, all samples kept)}

\item{t.filters}{An integer used to filter out rare taxa (default t.filters=1, taxa >= 1 occurrence kept)}

\item{iter}{An integer defining number of resampling iteration (default iter=0)}
}
\value{
A list containing the following components:
  \item{x}{DELTA S values for each live-dead comparisons (site-level differences in sample standardized
           species richness)}
  \item{y}{DELTA PIE values for each live-dead comparisons (site-level differences in evenness
           estimated as Hurlbert's PIE)}
  \item{z}{DELTA BETA Live-dead fidelity index for each live-dead comparison}
  \item{observed.means}{Grand means of diversity measures (means of  $x, $y, & $z) and
       by-group means when gp factor provided}
  \item{live}{The post-processed version of 'live' data matrix used in all analyses}
  \item{dead}{The post-processed version of 'dead' data matrix used in all analyses}
  \item{gp}{The post-processed version of 'gp' factor, when provided}
}
\description{
FidelityDiv provides estimates of differences in diversity between live and dead samples
using matrices (live and dead) with species/taxon abundance data. In the case of datastes
representing more than one sample, the function returns also means of differences.
If 'gp' factor is provided to aggregate sets of sites/samples, means for groups are returned as well.
}
\details{
FidelityDiv assess live-dead offsets in evenness/diversity using measures
of alpha diversity, alpha evenness, and beta diversity (returned as 3 separate objects):

(1) x - Live-dead offsets in alpha diversity for individual sites. The difference is measured as
ln(S)DEAD - ln(S)LIVE (i.e., difference between natural logarithms of sample-standardized
species richness of dead and live samples). A negative value indicates that alpha diversity
of live samples exceeds alpha diversity of dead sample (and vice versa).

(2) y - Live-dead offsets in evenness for individual sites. The difference is measured as
PIE(DEAD) - PIE(LIVE) (i.e., difference between estimates of Hurlbert's PIE for live and
dead samples). A negative value indicates that evenness of live samples exceeds evenness
of dead sample (and vice versa).

(3) z - Live-dead offsets in beta diversity for individual sites. The offset is estimated as
 Beta(DEAD) - Beta(LIVE). Three  measures of beta diversity can be selected:
 "beta variance", "Shannon's beta", "Whittaker's beta". A negative value indicates that
 beta diversity of live sample exceed beta diversity of dead sample (and vice versa).
}
\examples{

data(FidData)
out1 <- FidelityDiv(live = FidData$live, dead = FidData$dead, gp = FidData$habitat)


}
