% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelityEstimate.R
\name{FidelityEst}
\alias{FidelityEst}
\title{FidelityEst (computes live-dead fidelity measures)}
\usage{
FidelityEst(live, dead, gp = NULL, cor.measure = "spearman",
  dist.measure = "chao")
}
\arguments{
\item{live}{A matrix with counts of live-collected specimens (rows=samples, columns=species/taxa)}

\item{dead}{A matrix with counts of dead specimens (rows=samples, columns=species/taxa)
(dimensions must match)}

\item{gp}{An optional factor, with two or more levels, defining sample groups
(the length of gp must equal number of rows in live and dead)}
}
\value{
A list containing the following components:
  \item{x}{The values of the Spearman coefficients \emph{rho} for each live-dead comparison}
}
\description{
FidelityEst computes common live-dead fidelity measures for two datasets (live and dead)
}
\examples{

data(dune)
out1 <- FidelityEst(as.matrix(dune), as.matrix(dune[sample(1:nrow(dune)),sample(1:ncol(dune))]))
sapply(out1[1:2], mean)
plot(out1, xlim=c(-1,1), ylim=c(0,1), las=1, xlab=a$xlab, ylab=a$ylab); abline(h=0.5, v=0)

}
