% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelityEst.R
\name{FidelityEst}
\alias{FidelityEst}
\title{Compositional measures of live-dead fidelity}
\usage{
FidelityEst(live, dead, gp = NULL, cor.measure = "spearman",
  sim.measure = "chao", report = FALSE, n.filters = 0,
  t.filters = 1, iter = 0, dbzero = TRUE)
}
\arguments{
\item{live}{A matrix with counts of live-collected specimens (rows=sites, columns=taxa).
Dimensions of 'live' and 'dead' matrices must match exactely.}

\item{dead}{A matrix with counts of dead-collected specimens (rows=sites, columns=taxa).
Dimensions of 'live' and 'dead' matrices must match exactely.}

\item{gp}{An optional univariate factor defining groups of sites. The length of gp must
equal number of rows of 'live' and 'dead' matrices.}

\item{cor.measure}{A character string (default='spearman') defining correlation measure
(stats function 'vignettecor') used to estimate live-dead correlations.}

\item{sim.measure}{A character string (default='chao') defining similarity measure (vegan
function 'vegdist') used to estiamte live-dead similiarity. Any measure accepatble by
'vegdist' can be used.}

\item{report}{Logical (default=TRUE) to print compliance report from function 'FidelitySummary'.}

\item{n.filters}{An integer used to filter out small samples (default n.filters=0, all samples kept)}

\item{t.filters}{An integer used to filter out rare taxa (default t.filters=1, taxa >= 1 occurrence kept)}

\item{iter}{An integer defining number of resampling iteration (default iter=0)}

\item{dbzero}{Logical (default dbzero = TRUE) (removes double 0's when computing correlation measure)}
}
\value{
A list containing the following components:
  \item{x}{Live-dead correlation coefficients for each site}
  \item{y}{Live-dead similarity coefficients for each live-dead comparison}
  \item{z}{Live-dead fidelity index for each live-dead comparison}
  \item{measures}{The names of fidelity measures used}
  \item{observed.means}{Grand means of fidelity measures (means of  $x, $y, & $z) and
       by-group mean when 'gp' factor provided}
  \item{PF.stats}{Statistical summary under "Perfect Fidelity" model}
  \item{PF.dist}{Resampling distribution of fidelity measures under "Perfect Fidelity" model}
  \item{PF.dist.gp}{By-group resampling distribution of fidelity measures under
       "Perfect Fidelity" model when 'gp' factor provided}
  \item{gp.prob}{Pairwise significance tests for level combinations}
  \item{live}{The post-processed version of 'live' data matrix used in analyses}
  \item{dead}{The post-processed version of 'dead' data matrix used in analyses}
  \item{gp}{The post-processed version of 'gp', when 'gp' factor provided}
}
\description{
FidelityEst computes live-dead fidelity measures by comparing two matching matrices (live
and dead) with community abundance data. The function returns  fidelity estimates for indiviudal
sites and for groups of sites if 'gp' factor is provided.
}
\details{
FidelityEst assesses compositional fidelity using three measures
(returned as 3 separate objects):
(1) x - one of the three common measures of correlation/association:
 spearman (default), kendall, or pearson;
(2) y - an abundance based measure of similarity such as: jaccard-chao (default)
 or bray ('vegdist' in 'vegan');
(3) z - Fidelity Index measured as a scaled Euclidean distance from perfect fidelity score of
 indices x and y, where z = 0 denotes perfect fidelity (correlation and similarity measures
 both equal 1); z = 1 denotes no fidelity (correlation and similarity measures both equal 0);
 and z < 1 denotes cases where inverse correlation is observed.
}
\examples{

data(FidData)
out1 <- FidelityEst(live = FidData$live, dead = FidData$dead, gp = FidData$habitat)
SJPlot(out1, gpcol=c('forestgreen', 'coral3'))

}
