% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelityEst.R
\name{FidelityEst}
\alias{FidelityEst}
\title{FidelityEst (computes live-dead fidelity measures)}
\usage{
FidelityEst(live, dead, gp = NULL, cor.measure = "spearman",
  dist.measure = "chao")
}
\arguments{
\item{live}{A matrix with counts of live-collected specimens (rows=samples, columns=species/taxa)}

\item{dead}{A matrix with counts of dead specimens (rows=samples, columns=species/taxa)
(dimensions must match)}

\item{gp}{An optional factor, with two or more levels, defining sample groups
(the length of gp must equal number of rows in live and dead)}

\item{cor.measure}{Correlation coefficient to measure pairwise live-dead correlations (default='Spearman')}

\item{dist.measure}{Measure of similarity to measure live-dead agreement (default='Chao')}
}
\value{
A list containing the following components:
  \item{x}{The values of the Spearman coefficients \emph{rho} for each live-dead comparison}
}
\description{
FidelityEst computes common live-dead fidelity measures for two datasets (live and dead)
}
\examples{

library(vegan)
data(dune)
\dontrun{
out1 <- FidelityEst(as.matrix(dune), as.matrix(dune[sample(1:nrow(dune)),]))
plot(out1, xlim=c(-1,1), ylim=c(0,1), las=1, xlab=out1$xlab, ylab=out1$ylab)
abline(h=0.5, v=0)}
}
