% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelityEst.R
\name{FidelityEst}
\alias{FidelityEst}
\title{FidelityEst computes compositional measures of live-dead fidelity}
\usage{
FidelityEst(live, dead, gp = NULL, cor.measure = "spearman",
  sim.measure = "chao", report = FALSE, n.filters = 0,
  t.filters = 1, iter = 0, dbzero = TRUE)
}
\arguments{
\item{live}{A species/taxon abundance matrix with counts of live-collected specimens
(rows=samples, columns=species/taxa)}

\item{dead}{A matrix with counts of dead specimens (rows=samples, columns=species/taxa)
(dimensions of 'live' and 'dead' must match)}

\item{gp}{An optional factor, with two or more levels, defining sample groups
(the length of gp must equal number of rows in live and dead)}

\item{cor.measure}{defines correlation coefficient (stats function 'vignettecor') used to measure pairwise
live-dead correlations (default='spearman').}

\item{sim.measure}{A measure of similarity to measure live-dead agreement (default='Chao')}

\item{report}{Print compliance report from function FidelitySummary}

\item{n.filters}{An integer used to filter out small samples (default n.filters=0, all samples kept)}

\item{t.filters}{An integer used to filter out rare taxa (default t.filters=1, taxa >= 1 occurrence kept)}

\item{iter}{An integer defining number of resampling iteration (default iter=0)}

\item{dbzero}{Logical (default dbzero = TRUE) (removes double 0's when computing correlation measure)}
}
\value{
A list containing the following components:
  \item{x}{Live-dead correlation coefficients for each site}
  \item{y}{Live-dead similarity coefficients for each live-dead comparison}
  \item{z}{Live-dead fidelity index for each live-dead comparison}
  \item{measures}{The names of fidelity measures used}
  \item{observed.means}{Grand means of fidelity measures (means of  $x, $y, & $z) and
       by-group mean when 'gp' factor provided}
  \item{PF.stats}{Statistical summary under "Perfect Fidelity" model}
  \item{PF.dist}{Resampling distribution of fidelity measures under "Perfect Fidelity" model}
  \item{PF.dist.gp}{By-group resampling distribution of fidelity measures under
       "Perfect Fidelity" model when 'gp' factor provided}
  \item{gp.prob}{Pairwise significance tests for level combinations}
  \item{live}{The post-processed version of 'live' data matrix used in analyses}
  \item{dead}{The post-processed version of 'dead' data matrix used in analyses}
  \item{gp}{The post-processed version of 'gp', when 'gp' factor provided}
}
\description{
FidelityEst computes common live-dead fidelity measures by comparing two matching
matrices (live and dead) with species/taxon abundance data.
In the case of datastes representing more than one sample,
the function returns also means. If 'gp' factor is provided to define sample groups
means for groups are returned as well.
}
\details{
FidelityEst assess compositional fidelity using three measures
(returned as 3 separate objects):

(1) x - one of the three common measures of correlation/association:
 spearman (default), kendall, or pearson;

(2) y - an abundance based measure of similarity such as: jaccard-chao (default)
or bray ('vegdist' in 'vegan');

(3) z - Fidelity Index measured as a scaled Euclidean distance from perfect fidelity score of
 indices x and y, where z = 0 denotes perfect fidelity (correlation and similarity measures
 both equal 1); z = 1 denotes no fidelity (correlation and similarity measures both equal 0);
 and z < 1 denotes cases where inverse correlation is observed.
}
\examples{

data(FidData)
out1 <- FidelityEst(live = FidData$live, dead = FidData$dead, gp = FidData$habitat)
SJPlot(out1, gpcol=c('forestgreen', 'coral3'))

}
