% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelitySummary.R
\name{FidelitySummary}
\alias{FidelitySummary}
\title{FidelitySummary (to assess data compliance issues)}
\usage{
FidelitySummary(live, dead, gp = NULL, report = F)
}
\arguments{
\item{live}{A matrix with counts of live-collected specimens (rows=samples, columns=species/taxa)}

\item{dead}{A matrix with counts of dead specimens (rows=samples, columns=species/taxa)
(dimensions must match)}

\item{gp}{An optional factor, with two or more levels, defining sample groups
(the length of gp must equal number of rows in live and dead)}
}
\value{
Returns errors and critical warnings.
    No objects returned unless report=T to return warnings and data summary).
}
\description{
FidelitySummary function checks if the input objects are compliant.
This function is used invisibly in other PaleoFidelity functions. However, users can use it directly to
obtain a full list of warnings and a basic summary of user-provided objects (report=T)
}
\examples{

FidelitySummary(live=OBX$Live, dead=OBX$Dead, gp=)
}
