% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelitySummary.R
\name{FidelitySummary}
\alias{FidelitySummary}
\title{FidelitySummary (to assess data compliance issues)}
\usage{
FidelitySummary(live, dead, gp = NULL, report = FALSE, n.filters = 0,
  t.filters = 1)
}
\arguments{
\item{live}{A matrix with counts of live-collected specimens (rows=samples, columns=species/taxa)
(dimensions must match that of 'live' object)}

\item{dead}{A matrix with counts of dead specimens (rows=samples, columns=species/taxa)
(dimensions must match that of 'live' object)}

\item{gp}{An optional factor, with two or more levels, defining sample groups.
The length of gp must equal number of rows in live and dead and at least two levels must
include 2 or more observations to allow for by-group analyses.}

\item{report}{Logical (default=FALSE), set report=TRUE to print additional warnings and data summary}

\item{n.filters}{Integer (default = 0) to remove small samples with n < n.filters occurrences}

\item{t.filters}{Integer (default = 1) to remove rare taxa with t < t.filters occurrences}
}
\value{
A list including the following components:
  \item{live}{The filtered live dataset where rows=sites and columns=taxa}
  \item{dead}{The filtered dead dataset where rows=sites and columns=taxa}
  \item{gp}{The grouping factor associated with sites (if provided)}
}
\description{
FidelitySummary function checks if the input objects are appropriately formatted and contain
adequate data needed to carry out fidelity analysis. The function also allows the user to
filter data by removing small samples and rare taxa
}
\details{
This function is implemented in other PaleoFidelity functions. However, users
are encouraged to use it prior to any other functions in order to:
1). Check for errors and warnings;
2). Generate a basic summary of user-provided objects (set logical argument report=TRUE)
3). Explore how data filtering (arguments "n.filters" and "t.filters") affect data dimensionality

NOTE: FidelitySummary function will provide an initial compliance evaluation.
Also the function allows the user to detect small samples and assess
if removing those samples is advisable in terms of data dimensionality.
If removal of small samples is desirable for subsequent analyses,
please specify the numerical value of "n.filters" and "t.filters" in other PaleoFidelity functions.
}
\examples{
data(FidData)
FidelitySummary(live=FidData$live, dead=FidData$dead, report=TRUE)
FidelitySummary(live=FidData$live, dead=FidData$dead, gp=FidData$habitat, report=TRUE, n.filters=30)

}
