% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelitySummary.R
\name{FidelitySummary}
\alias{FidelitySummary}
\title{FidelitySummary (to assess data compliance issues)}
\usage{
FidelitySummary(live, dead, gp = NULL, report = FALSE, n.filters = 0)
}
\arguments{
\item{live}{A matrix with counts of live-collected specimens (rows=samples, columns=species/taxa)}

\item{dead}{A matrix with counts of dead specimens (rows=samples, columns=species/taxa)
(dimensions must match)}

\item{gp}{An optional factor, with two or more levels, defining sample groups.
The length of gp must equal number of rows in live and dead and at least two levels must
include 2 or more observations to allow for by-group analyses.}

\item{report}{Logical (default=FALSE), set report=TRUE to print additional warnings and data summary}

\item{n.filters}{Numerical value used to remove small samples with n < n.filters (default value is 0)}
}
\value{
Returns errors and critical warnings.
    No objects returned unless report=TRUE (return additional warnings and data summary)
}
\description{
FidelitySummary function checks if the input objects are compliant.
This function is used invisibly in other PaleoFidelity functions. However, users can use it directly to
obtain a full list of warnings and a basic summary of user-provided objects (report=T)
}
\examples{
data(FidData)
FidelitySummary(live=FidData$live, dead=FidData$dead, report=TRUE)
FidelitySummary(live=FidData$live, dead=FidData$dead, gp=FidData$habitat, report=TRUE, n.filters=30)

}
