% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FidelitySummary.R
\name{FidelitySummary}
\alias{FidelitySummary}
\title{FidelitySummary (to assess data compliance issues)}
\usage{
FidelitySummary(live, dead, gp = NULL, report = FALSE, n.filters = 0)
}
\arguments{
\item{live}{A matrix with counts of live-collected specimens (rows=samples, columns=species/taxa)}

\item{dead}{A matrix with counts of dead specimens (rows=samples, columns=species/taxa)
(dimensions must match)}

\item{gp}{An optional factor, with two or more levels, defining sample groups.
The length of gp must equal number of rows in live and dead and at least two levels must
include 2 or more observations to allow for by-group analyses.}

\item{report}{Logical (default=FALSE), set report=TRUE to print additional warnings and data summary}

\item{n.filters}{Numerical value used to remove small samples with n < n.filters (default value is 0)}
}
\value{
When report=TRUE, the function returns a matrix with data summary and may also return warnings.
    When report=FALSE, no objects are returned (only warnings may be produced)
}
\description{
FidelitySummary function checks if the input objects are appropriately formatted and contain
adequate data needed to carry out a mathematically meaningful fidelity analysis.
}
\details{
This function is implemented in other PaleoFidelity functions. However, users are encouraged to use
it prior to any other functioons in order to:
1). Check for errors and warnings;
2). Generate a basic summary of user-provided objects (make sure to set parameter report=TRUE)
3). Explore how data filtering (parameter "n.filters") affect the dimensionality of the data.

NOTE: FidelitySummary function will not alter data objects passed through it, but only provide an initial
compliance evaluation. Also the function allows the user to detect small samples and assess if removing
those samples is advisable in terms of data dimensionality. If removal of small samples is desirable
for subsequent analyses, please specify the numerical value of "n.filters" in other PaleoFidelity functions.
}
\examples{
data(FidData)
FidelitySummary(live=FidData$live, dead=FidData$dead, report=TRUE)
FidelitySummary(live=FidData$live, dead=FidData$dead, gp=FidData$habitat, report=TRUE, n.filters=30)

}
