% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDPlot.R
\name{LDPlot}
\alias{LDPlot}
\title{A Comparative Live-Dead Barplot}
\usage{
LDPlot(live, dead, tax.names, toplimit = 10, barwidth = 150/toplimit,
  col1 = "black", col2 = "gray", arr.col = "red", arr.lty = 1,
  cex.axis = 0.7, tck = -0.02, cex.lab = 0.8, cex.names = 0.95,
  cex.label = 1)
}
\arguments{
\item{live}{A vector of integers with counts of live specimens by taxa (or other units)}

\item{dead}{A vector of integers with counts of dead specimens by taxa (or other units)}

\item{tax.names}{A vector with a list of names of taxa (or other units
used as variables)}

\item{toplimit}{A numerical value (default = 10) defining the number of top
species (or other variables) to be plotted}

\item{barwidth}{A numerical value (default = 250 / toplimit) defining bar width
(more precisely, the thickness of \code{\link[graphics]{lines}} used to
represent bars)}

\item{col1}{A character string (default = 'black') defining the color of bars
for taxa (or other variables) shared by live and dead data}

\item{col2}{A character string (default = 'gray') defining the color of bars
for taxa (or other variables) unique to either live or dead data}

\item{arr.col}{A character string (default = 'red') defining the color of arrows}

\item{arr.lty}{An integer or character string specifying type of line used for
arrows (default = 1) (this value is passed on to the graphical parameter 'lty')}

\item{cex.axis}{A numerical value (default = 0.7).The magnification to be used for
x axis annotation relative to the current setting of cex.}

\item{tck}{A numerical value (default  = -0.02) defining length of tick marks
for x axis}

\item{cex.lab}{A numerical value (default = 0.8) defining font size for
x-axis label}

\item{cex.names}{A numerical value (default = 0.95) defining font size
for names of taxa/variables corresponding to individual bars on the barplot}

\item{cex.label}{A numerical value (defult = 1) defining font size for 'Live'
and 'Dead' titles placed above the barplots}
}
\value{
A single plot.
}
\description{
LDPlot function generates a comparative live-dead plot of the most frequent taxa
(or other variables) to highlight live-dead congruences and live-dead discordances
in relative abundance and rank order of the most abundant taxa/variables.
}
\details{
LDPlot function produces a plot comparing barplots of the top "n"
most common taxa observed in live and dead datasets
(e.g., Kowalewski et al. 2003). These L-D comparisons can apply to
single sites or data pooled across multiple sites. Because taxa/variable
names can vary in length and because the number of plotted taxa/variables
can span a wide range of values, plot margins and cex paramaters may need
adjustments.
}
\examples{

temp.par <- par(mar=c(3,6,1,6))
LDPlot(live=colSums(FidData$live), dead=colSums(FidData$dead),
tax.names=colnames(FidData$live), toplimit=15, barwidth = 21,
col1 = 'green2', col2 = 'red4', arr.col = 'green2', arr.lty = 1)
par(temp.par)

}
\references{
Kowalewski, M.,	Carroll, M., Casazza, L., Gupta, N., Hannisdal, B.,
Hendy, A., Krause, R.A., Jr., Labarbera, M., Lazo, D.G., Messina, C., Puchalski, S.,
Rothfus, T.A., Sälgeback, J., Stempien, J., Terry, R.C., Tomašových, A., (2003),
Quantitative fidelity of brachiopod-mollusk assemblages from modern subtidal
environments of San Juan Islands, USA. Journal of Taphonomy 1: 43-65.
}
