% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SJPlot.R
\name{SJPlot}
\alias{SJPlot}
\title{A Fidelity Plot}
\usage{
SJPlot(
  x,
  bubble = TRUE,
  xlim = c(-1, 1),
  ylim = c(0, 1),
  trans = 0.3,
  cex = 1,
  cex.mean = 2,
  cex.model = 1,
  cex.model.mean = 1,
  cex.bubble = 2,
  cex.legend = 0.8,
  axes = T,
  pch = 21,
  col = "black",
  col.model = "gray",
  col.mean = "white",
  col.model.mean = "white",
  gpcol = NULL,
  pch2 = 21,
  pch3 = ".",
  pch4 = 21,
  PF = FALSE,
  CI = TRUE,
  adjF = TRUE,
  ssF = FALSE,
  unadjF = FALSE,
  addlegend = FALSE,
  addinfo = FALSE,
  addbubble = FALSE,
  info.y = 0.5,
  bubble.y = -0.15,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{x}{An object of the class 'FidelityEst' returned by \code{\link{FidelityEst}} function.}

\item{bubble}{Logical (default=TRUE): to produce a bubble plot with symbols scaled by N-min
(the number of observations in the smaller of the two (live vs. dead) compared samples).}

\item{xlim}{A vector with two numerical values defining x axis limits (default = c(-1, 1)).}

\item{ylim}{A vector with two numerical values defining y axis limits (default = c(0, 1)).}

\item{trans}{A numerical value (default = 0.3) defining transparency of background fill for symbols.}

\item{cex}{A numerical value (default = 1) defining symbol size (applicable if Bubble = FALSE).}

\item{cex.mean}{A numerical value (default = 2) defining symbol size for means
across all samples or sample groups.}

\item{cex.model}{A numerical value (default = 1) defining symbol size for
individual model estimates.}

\item{cex.model.mean}{A numerical value (default = 1) defining symbol size for
mean sample model estimates.}

\item{cex.bubble}{A numerical value (default = 2) defining scaling factor for bubbles.}

\item{cex.legend}{A numerical value (default = 1) defining text size for legend.}

\item{axes}{Logical (default = TRUE): to determine if axes should be plotted}

\item{pch}{An integer or a single character (default = 21) specifying symbol type.}

\item{col}{A character string (default = 'black') defining symbol color.}

\item{col.model}{A character string (default = 'gray') defining symbol color for model points.}

\item{col.mean}{A character string (default = 'black') defining symbol color for the mean.}

\item{col.model.mean}{A character string (default = 'white') defining symbol color
for model sample means.}

\item{gpcol}{Color names (default=1:length(levels(x$gp))) defining colors for sample groups,
applicable when 'gp' factor was provided in \code{\link{FidelityEst}} function.
If custom colors are provided, the number of colors must match number of levels
in a 'gp' factor.}

\item{pch2}{An integer or a single character (default=21) specifying symbol type
for grand mean or group means.}

\item{pch3}{An integer or a single character (default='.') specifying symbol type
for simulated sample values.}

\item{pch4}{An integer or a single character (default=21) specifying symbol type
for simulated mean sample values.}

\item{PF}{Logical (default=FALSE): adds a scatter plot of resampled fidelity
estimates for the null model of perfect fidelity.}

\item{CI}{Logical (default=TRUE): adds confidence intervals to corrected fidelity estimates}

\item{adjF}{Logical (default=TRUE): plots adjusted fidelity estimates}

\item{ssF}{Logical (default=FALSE): plots sample-standardized fidelity estimates}

\item{unadjF}{Logical (default=FALSE): plots non-adjusted fidelity estimates}

\item{addlegend}{Logical (default=FALSE): adds legend to the plot, if 'gp' factor is provided.}

\item{addinfo}{Logical (default=FALSE): prints parameter values above the plot}

\item{addbubble}{Logical (default=FALSE): prints legend relating bubble sizes to
N-min, which denotes the number of specimens in the smaller of the two compared
samples (usually, 'live' sample is the smaller of the two)}

\item{info.y}{Numeric (default = 0.5): specifies how far above the plot should
"addinfo" legend (see above) be plotted. This is "line" parameter value
in \code{\link[graphics]{mtext}} function)}

\item{bubble.y}{Numeric (default = - 0.15): specifies how far above the plot should
"addbubble" legend (see above) be plotted. This is "inset" parameter value
in \code{\link[graphics]{legend}} function)}

\item{xlab}{Character string that provides x-axis label. The default value is inherited
from FidelityEst object based on a correlation measure used there.}

\item{ylab}{Character string that provides y-axis label. The default value is inherited
from FidelityEst object based on a similarity measure used there.}
}
\value{
A single bivariate plot.
}
\description{
SJPlot function produces a bivariate plot of a correlation measure (x axis)
and similarity measure (y axis) to depict live-dead fidelity at one or more
sites.
}
\details{
SJPlot function uses the object of the class 'FidelityEst', produced by
\code{\link{FidelityEst}} function, to generate a live-dead fidelity plot.
If default arguments for fidelity measures are used in \code{\link{FidelityEst}} function,
a Spearman vs. Jaccard-Chao fidelity plot (as in Kidwell, 2007) is produced.
If a grouping factor is provided, symbols are color-coded by levels and group means
are plotted. Bivariate distributions produced by a resampling model can be included.

NOTE: This function utilizes \code{\link[graphics]{plot.default}} function,
including some of its arguments. It allows to explore visually various alternative
estimates of fidelity.
}
\examples{

out1 <- FidelityEst(FidData$live[6:9,], FidData$dead[6:9,], FidData$habitat[6:9],
                   n.filters=30, iter=100, t.filters=1)
SJPlot(out1)

}
\references{
Kidwell, S.M., 2007, Discordance between living and death assemblages
as evidence for anthropogenic ecological change. Proc Natl Acad Sci USA 104(45): 17701–17706.
}
